issues/383: Check new SocketImpl and SocketChannel methods added in newer versions of Java
